MeetTrack-main/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.50
botocore==1.42.50
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.3
packaging==26.0
pydantic==2.12.5
pydantic-core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.2
s3transfer==0.16.0
six==1.17.0
smart-open==7.5.0
SQLAlchemy==2.0.46
starlette==0.52.1
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
typing-extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
google-generativeai==0.3.1
